+++ common-reducers/src/test/java/com/mattunderscore/trees/reducers/LongSummingReducerTest.java
@@ -1,0 +1,70 @@
+/* Copyright © 2016 Matthew Champion
+All rights reserved.
+
+Redistribution and use in source and binary forms, with or without
+modification, are permitted provided that the following conditions are met:
+ * Redistributions of source code must retain the above copyright
+      notice, this list of conditions and the following disclaimer.
+ * Redistributions in binary form must reproduce the above copyright
+      notice, this list of conditions and the following disclaimer in the
+      documentation and/or other materials provided with the distribution.
+ * Neither the name of mattunderscore.com nor the
+      names of its contributors may be used to endorse or promote products
+      derived from this software without specific prior written permission.
+
+THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND
+ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED
+WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE ARE
+DISCLAIMED. IN NO EVENT SHALL MATTHEW CHAMPION BE LIABLE FOR ANY
+DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES
+(INCLUDING, BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES;
+LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND
+ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT
+(INCLUDING NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS
+SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE. */
+
+package com.mattunderscore.trees.reducers;
+
+import com.mattunderscore.trees.Trees;
+import com.mattunderscore.trees.construction.BottomUpTreeBuilder;
+import com.mattunderscore.trees.immutable.TreeNodeImpl;
+import com.mattunderscore.trees.tree.Node;
+import com.mattunderscore.trees.tree.Tree;
+import org.junit.Before;
+import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Unit tests for {@link LongSummingReducer}.
+ *
+ * @author Matt Champion on 31/08/2015
+ */
+public final class LongSummingReducerTest {
+    private static final Trees TREES = Trees.get();
+    private Tree<Long, Node<Long>> tree;
+
+    @Before
+    public void setUp() {
+        final BottomUpTreeBuilder<Long, Node<Long>> builder = TREES.treeBuilders().bottomUpBuilder();
+        tree = builder.create(
+            5L,
+            builder.create(3L),
+            builder.create(
+                3L,
+                builder.create(
+                    7L,
+                    builder.create(9L))))
+            .build(TreeNodeImpl.<Long>typeKey());
+    }
+
+    @Test
+    public void sum() {
+        final LongSummingReducer reducer = LongSummingReducer.get();
+        final long result = TREES
+                .querier()
+                .reduce(tree, reducer);
+
+        assertEquals(27L, result);
+    }
+}