+++ CHANGELOG.md
@@ -129,3 +129,4 @@ source nodes
 * Added support for querying if a tree is balanced 
 * Added additional matchers based on comparators
 * Added support for applying a reducing function in post-order to a tree or subtree
+* Added module for common reducers